Migrate segArea code to its own file.
